cfg-setters: packaging & code-style refacto
