+++ internal/setters/parsers/builder_parser.go
@@ -1,0 +1,5 @@
+package parsers
+
+type Parser interface {
+	parse(string, bool) (interface{}, error)
+}
