+++ internal/setters/parsers/string_parser.go
@@ -1,0 +1,12 @@
+package parsers
+
+type stringParser struct{}
+
+func (*stringParser) parse(v string, ptr bool) (interface{}, error) {
+	if ptr {
+		x := v
+		return &x, nil
+	}
+
+	return v, nil
+}
